+++ Gemfile
@@ -35,7 +35,7 @@ gem "json"
 gem 'rest-client'
 
 # Algolia (Search Function)
-gem 'algolia'
+gem "algoliasearch-rails"
 
 # Use Redis adapter to run Action Cable in production
 # gem "redis", "~> 4.0"
