+++ app/views/portfolios/_portfolio-form.html.erb
@@ -2,8 +2,7 @@
     <div class="xl:w-70">
       <div class="input-group relative items-stretch w-full">
         <%= form.input :name %>
-        <%= form.submit
-         %>
+        <%= form.submit %>
       </div>
     </div>
   <% end %>
