+++ app/views/stocks/_search_bar.html.erb
